WIP - both chatbots look right?
