+++ chat-ui/src/constants.ts
@@ -1,2 +1,12 @@
 // The maximum number of characters allowed in a single user message.
 export const MAX_INPUT_CHARACTERS = 300;
+
+export const WELCOME_MESSAGE =
+  "Welcome to MongoDB AI Assistant. What can I help you with?";
+
+export const SUGGESTED_PROMPTS = [
+  "How do you deploy a free cluster in Atlas?",
+  "How do you import or migrate data into MongoDB Atlas?",
+  "How do I get started with MongoDB?",
+  "Why should I use Atlas Search?",
+];
